+++ src/coll_input.h
@@ -1,6 +1,2 @@
-/*void version(void);
-
-void help(void);
-
 void (*coll_args (int argc, char **argv,maze_t* maze,tile_set_t** tile_set, FILE** output_file))(maze_t*);
-*/
+
