--- test/test_collin.c
@@ -1,1 +1,0 @@
-Hello World
